x trying to find and correct bugs detected by the compiler and/or Coverty
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@6321 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/File__Analyze.cpp
@@ -3376,7 +3376,7 @@ bool File__Analyze::Demux_UnpacketizeContainer_Test_OneFramePerFile ()
     {
         size_t* File_Buffer_Size_Hint_Pointer=Config->File_Buffer_Size_Hint_Pointer_Get();
         if (File_Buffer_Size_Hint_Pointer)
-            (*File_Buffer_Size_Hint_Pointer)=Config->File_Current_Size-Config->File_Current_Offset-Buffer_Size;
+            (*File_Buffer_Size_Hint_Pointer) = (size_t)(Config->File_Current_Size - Config->File_Current_Offset - Buffer_Size);
         return false;
     }
