+++ Source/MediaInfo/File__Analyze_Buffer_MinimizeSize.cpp
@@ -1210,7 +1210,7 @@ void File__Analyze::Get_ISO_6937_2(int64u Bytes, Ztring &Info)
 {
     INTEGRITY_SIZE_ATLEAST_STRING(Bytes);
     Info.clear();
-    size_t End=Buffer_Offset+(size_t)Element_Offset+Bytes;
+    size_t End = Buffer_Offset + (size_t)Element_Offset + (size_t)Bytes;
     for (size_t Pos=Buffer_Offset+(size_t)Element_Offset; Pos<End; ++Pos)
     {
         wchar_t EscapeChar=__T('\x0000');
@@ -1332,7 +1332,7 @@ void File__Analyze::Get_ISO_8859_5(int64u Bytes, Ztring &Info)
 {
     INTEGRITY_SIZE_ATLEAST_STRING(Bytes);
     Info.clear();
-    size_t End=Buffer_Offset+(size_t)Element_Offset+Bytes;
+    size_t End = Buffer_Offset + (size_t)Element_Offset + (size_t)Bytes;
     for (size_t Pos=Buffer_Offset+(size_t)Element_Offset; Pos<End; ++Pos)
     {
         switch (Buffer[Pos])
